fs: sdfat: Update to driver version 2.0.6
 * From Samsung package version: G960USQU1ARBG

+++ fs/sdfat/api.h
@@ -119,13 +119,13 @@ typedef struct {
 } DATE_TIME_T;
 
 typedef struct {
-	u32      Offset;    // start sector number of the partition
-	u32      Size;      // in sectors
+	u64      Offset;    // start sector number of the partition
+	u64      Size;      // in sectors
 } PART_INFO_T;
 
 typedef struct {
 	u32      SecSize;    // sector size in bytes
-	u32      DevSize;    // block device size in sectors
+	u64      DevSize;    // block device size in sectors
 } DEV_INFO_T;
 
 typedef struct {
@@ -139,7 +139,7 @@ typedef struct {
 /* directory structure */
 typedef struct {
 	u32      dir;
-	s32      size;
+	u32      size;
 	u8       flags;
 } CHAIN_T;
 
@@ -147,7 +147,7 @@ typedef struct {
 typedef struct {
 	u32      clu;
 	union {
-		s32 off;     // cluster offset
+		u32 off;     // cluster offset
 		s32 eidx;    // entry index
 	};
 } HINT_T;
@@ -209,7 +209,7 @@ typedef struct __cache_entry {
 		struct __cache_entry *next;
 		struct __cache_entry *prev;
 	} hash;
-	u32 sec;
+	u64 sec;
 	u32 flag;
 	struct buffer_head   *bh;
 } cache_ent_t;
@@ -223,7 +223,7 @@ typedef struct __FATENT_OPS_T {
 } FATENT_OPS_T;
 
 typedef struct {
-	s32      (*alloc_cluster)(struct super_block *, s32, CHAIN_T *, int);
+	s32      (*alloc_cluster)(struct super_block *, u32, CHAIN_T *, s32);
 	s32      (*free_cluster)(struct super_block *, CHAIN_T *, s32);
 	s32      (*count_used_clusters)(struct super_block *, u32 *);
 	s32      (*init_dir_entry)(struct super_block *, CHAIN_T *, s32, u32, u32, u64);
@@ -253,16 +253,16 @@ typedef struct __FS_INFO_T {
 	s32	 bd_opened;              // opened or not
 	u32      vol_type;               // volume FAT type
 	u32      vol_id;                 // volume serial number
-	u32      num_sectors;            // num of sectors in volume
+	u64      num_sectors;            // num of sectors in volume
 	u32      num_clusters;           // num of clusters in volume
 	u32      cluster_size;           // cluster size in bytes
 	u32      cluster_size_bits;
 	u32      sect_per_clus;        // cluster size in sectors
 	u32      sect_per_clus_bits;
-	u32      FAT1_start_sector;      // FAT1 start sector
-	u32      FAT2_start_sector;      // FAT2 start sector
-	u32      root_start_sector;      // root dir start sector
-	u32      data_start_sector;      // data area start sector
+	u64      FAT1_start_sector;      // FAT1 start sector
+	u64      FAT2_start_sector;      // FAT2 start sector
+	u64      root_start_sector;      // root dir start sector
+	u64      data_start_sector;      // data area start sector
 	u32      num_FAT_sectors;        // num of FAT sectors
 	u32      root_dir;               // root dir cluster
 	u32      dentries_in_root;       // num of dentries in root dir
@@ -337,7 +337,7 @@ s32 fsapi_rename(struct inode *old_parent_inode, FILE_ID_T *fid,
 s32 fsapi_unlink(struct inode *inode, FILE_ID_T *fid);
 s32 fsapi_read_inode(struct inode *inode, DIR_ENTRY_T *info);
 s32 fsapi_write_inode(struct inode *inode, DIR_ENTRY_T *info, int sync);
-s32 fsapi_map_clus(struct inode *inode, s32 clu_offset, u32 *clu, int dest);
+s32 fsapi_map_clus(struct inode *inode, u32 clu_offset, u32 *clu, int dest);
 s32 fsapi_reserve_clus(struct inode *inode);
 
 /* directory management functions */
